[opentitanlib] Move num_de from otp to util
The num_de module that is used for deserializing certain numeric types
from HJSON files belongs in util. Move it there so it can also be used
in the signer.
