+++ sw/host/opentitanlib/BUILD
@@ -32,7 +32,6 @@ rust_library(
         "src/lib.rs",
         "src/otp/lc_state.rs",
         "src/otp/mod.rs",
-        "src/otp/num_de.rs",
         "src/otp/otp.rs",
         "src/otp/otp_img.rs",
         "src/otp/otp_mmap.rs",
@@ -67,6 +66,7 @@ rust_library(
         "src/util/file.rs",
         "src/util/image.rs",
         "src/util/mod.rs",
+        "src/util/num_de.rs",
         "src/util/parse_int.rs",
         "src/util/present.rs",
         "src/util/usb.rs",
