+++ sw/host/opentitanlib/src/otp/mod.rs
@@ -8,8 +8,6 @@ pub mod otp_img;
 pub mod otp_mmap;
 pub mod vmem_serialize;
 
-mod num_de;
-
 #[cfg(test)]
 mod tests {
     use super::*;
