+++ sw/host/opentitanlib/src/util/num_de.rs
